# docs/swagger/collections.yaml
openapi: 3.0.0
servers:
  - description: Listmonk Developement Server
    url: http://localhost:9000/api
info:
  version: "1.0.0"
  title: Listmonk
  description: The API collection for listmonk
  license:
    name: AGPL-3.0 license
    url: https://github.com/knadh/listmonk/blob/master/LICENSE
...
paths:
...
  /bounces:
    get:
      description: handles retrieval of bounce records.
      operationId: getBounces
      tags:
        - Bounces
      parameters:
        - in: query
          name: campaign_id
          description: Numeric identifier for retrieving bounce records associated with a specific campaign
          schema:
            type: integer
        - in: query
          name: page
          description: Page number for paginated results. Start from 1 for the first page
          schema:
            type: integer
        - in: query
          name: per_page
          description: Number of items per page. Use an integer for specific page size or 'all' to retrieve all results
          schema:
            oneOf:
              - type: integer
                description: Number of items to return per page
              - type: string
                enum:
                  - "all"
                description: Return all results without pagination
        - in: query
          name: source
          description: Filter bounce records by their source of origin
          schema:
            type: string
        - in: query
          name: order_by
          description: Specifies the field by which to sort the bounce records. Available options are 'email', 'campaign_name', 'source', and 'created_at'
          schema:
            type: string
            enum: ["email", "campaign_name", "source", "created_at"]
        - in: query
          name: order
          description: Determines the sort order of results. Use 'asc' for ascending or 'desc' for descending order
          schema:
            type: string
            enum: ["asc", "desc"]
      responses:
        "200":
          description: list of bounce records
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      results:
                        type: array
                        items:
                          $ref: "#/components/schemas/Bounce"
                      query:
                        type: string
                      total:
                        type: integer
                      per_page:
                        type: integer
                      page:
                        type: integer
    delete:
      description: handles deletion of bounce records.
      operationId: deleteBounces
      tags:
        - Bounces
      parameters:
        - in: query
          name: all
          description: flag for multiple bounce record deletion
          schema:
            type: boolean
        - in: query
          name: id
          description: list of bounce ids to delete
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: boolean
  "/bounces/{id}":
    get:
      description: handles retrieval of bounce record by id
      operationId: getBounceById
      parameters:
        - in: path
          name: id
          required: true
          description: The id value of the bounce you want to retreive.
          schema:
            type: integer
      tags:
        - Bounces
      responses:
        "200":
          description: bounce object
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Bounce"

    delete:
      description: handles bounce deletion, either a single one (ID in the URI), or a list.
      operationId: deleteBounceById
      parameters:
        - in: path
          name: id
          required: true
          description: The id value of the bounce you want to delete.
          schema:
            type: integer
      tags:
        - Bounces
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: boolean
...
components:
  schemas:
...
    Bounce:
      type: object
      properties:
        results:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              type:
                type: string
              source:
                type: string
              meta:
                type: object
              created_at:
                type: string
              email:
                type: string
              subscriber_uuid:
                type: string
              subscriber_id:
                type: integer
              campaign:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
              campaign_uuid:
                type: string
              total:
                type: integer
...